게시글 CRUD 인증, 인가 코드 추가

+++ src/main/java/com/sparta/newsfeedproject/service/PostService.java
@@ -3,12 +3,14 @@ package com.sparta.newsfeedproject.service;
 import com.sparta.newsfeedproject.dto.PostDto;
 import com.sparta.newsfeedproject.entity.PostEntity;
 import com.sparta.newsfeedproject.repository.PostRepository;
+import com.sparta.newsfeedproject.security.UserDetailsImpl;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
@@ -17,8 +19,12 @@ public class PostService {
     private final PostRepository postRepository;
 
     // 게시글 작성
-    public void save(PostDto postDto) {
-        PostEntity postEntity = PostEntity.toSaveEntity(postDto);
+    public void save(PostDto postDto, UserDetailsImpl userDetails) {
+//        PostEntity postEntity = PostEntity.toSaveEntity(postDto);
+//        postEntity.setPostWriter(userDetails.getUsername());
+
+        PostEntity postEntity = new PostEntity(postDto);
+        postEntity.setPostWriter(userDetails.getUsername());
         postRepository.save(postEntity);
     }
 
@@ -49,14 +55,30 @@ public class PostService {
         }
     }
 
+    // 게시글 수정
+    @Transactional
+    public PostDto update(PostDto postDto, Long id, UserDetailsImpl userDetails) {
+//        PostEntity postEntity = PostEntity.toUpdateEntity(postDto);
+//        postRepository.save(postEntity);
+        PostEntity postEntity = postRepository.findById(id).orElseThrow(() -> new NoSuchElementException("해당 게시글이 존재하지 않습니다."));
 
-    public PostDto update(PostDto postDto) {
-        PostEntity postEntity = PostEntity.toUpdateEntity(postDto);
-        postRepository.save(postEntity);
+        if (postEntity.getPostWriter().equals(userDetails.getUsername())) {
+            postEntity.setTitle(postDto.getTitle());
+            postEntity.setContents(postDto.getContents());
+        } else {
+            throw new IllegalArgumentException("아이디가 일치하지 않습니다.");
+        }
+
         return findById(postDto.getId());
     }
 
-    public void delete(Long id) {
-        postRepository.deleteById(id);
+    // 게시글 삭제
+    public void delete(Long id, UserDetailsImpl userDetails) {
+        PostEntity postEntity = postRepository.findById(id).orElseThrow(() -> new NoSuchElementException("해당 게시글이 존재하지 않습니다."));
+        if(postEntity.getPostWriter().equals(userDetails.getUsername())) {
+            postRepository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException("아이디가 일치하지 않습니다.");
+        }
     }
 }
